Tidy up basic_headers for documentation
